source/04_rag/06_rag_generation/requirements.txt
chromadb
langchain-core
openai
python-dotenv
